Split into 2 components
Now Form and StockList are two components pass into top level App componenet
